refactor TableOutput into component
commit tobe rebased
